+++ src/plugin_parser/__init__.py
@@ -1,3 +1,5 @@
 """
 Copyright (c) Cutleast
 """
+
+from .plugin_parser import PluginParser